#3963 Moving the contents of V1_2-branch/CA-Certificates-Base to this new top-level svn directory.
I also added a new defs file, and edited all the scripts to remove
the reliance on make-vdt, as well as some minor code cleanup / bug
removal.

This area is now organized so that VDT specific scripts are in
vdt-scripts, so that we can share this part of the repository with
the general public and they can use our scripts to help maintain
a certificate distribution.

As of this checkin, a number of pieces of code are commented out
to prevent accidental over-writing of production components on
our web server.  When we first update the certs, we will need to
fix these.  I marked them all with a comment that contains "EDIT"


git-svn-id: https://vdt.cs.wisc.edu/svn/certs/trunk@8032 4e558342-562e-0410-864c-e07659590f8c